fix: reuse fit tabs in screenshot export
